fix detail
