+++ app/views/jobs/_job.html.erb
@@ -1,5 +1,38 @@
 <%= button_to job_path(job), class: "mt-2", method: :get do %>
   <div class="job">
-    <%= render 'job_content', job: job, user: job.user %>
+    <!-- レンダリングの回数を減らすためパーシャル化解除 -->
+    <div class="job-content">
+      <h3><%= job.title %></h3>
+      <div class="card mt-4">
+        <div class="card-header">
+          <strong>発注者</strong>
+        </div>
+        <div class="card-body user-box">
+          <% user = job.user %>
+          <%= image_tag user.show_image, class: "mini-user-image" if user.image.attached? %>
+          <p><%= link_to user.name, user_path(user), class: "ml-2" %></p>
+        </div>
+      </div>
+      <div class="card mt-2">
+        <div class="card-header">
+          <strong>報酬</strong>
+        </div>
+        <div class="card-body">
+          <% if job.reward_type.present? %>
+            <p><%= show_reward(job) %></p>
+          <% else %>
+            <p><%= job.pay %></p>
+          <% end %>
+        </div>
+      </div>
+      <div class="card mt-2">
+        <div class="card-header">
+          <strong>説明</strong>
+        </div>
+        <div class="card-body">
+          <p><%= safe_join(job.explanation.split("\n"), tag(:br)) %></p>
+        </div>
+      </div>
+    </div>
   </div>
 <% end %>
