+++ app/views/jobs/_job_content.html.erb
@@ -4,8 +4,8 @@
     <div class="card-header">
       <strong>発注者</strong>
     </div>
-    <div class="card-body">
-      <p><%= link_to job.user.name, user_path(job.user) %></p>
+    <div class="card-body user-box">
+      <%= render 'shared/mini_user', user: job.user %>
     </div>
   </div>
   <div class="card mt-2">
