+++ app/views/messages/show.html.erb
@@ -4,16 +4,16 @@
     <div class="card-header">
       <strong>送信者</strong>
     </div>
-    <div class="card-body">
-      <p><%= link_to @message.sender.name, user_path(@message.sender) %></p>
+    <div class="card-body user-box">
+      <%= render 'shared/mini_user', user: @message.sender %>
     </div>
   </div>
   <div class="card mt-2">
     <div class="card-header">
       <strong>受信者</strong>
     </div>
-    <div class="card-body">
-      <p><%= link_to @message.receiver.name, user_path(@message.receiver) %></p>
+    <div class="card-body user-box">
+      <%= render 'shared/mini_user', user: @message.receiver %>
     </div>
   </div>
   <div class="card mt-2">
@@ -22,7 +22,7 @@
     </div>
     <div class="card-body">
       <% if text = @message.content.presence %>
-        <p><%= safe_join(text.split("\n"), tag(:br)) %><p>
+        <p><%= safe_join(text.split("\n"), tag(:br)) %></p>
       <% end %>
     </div>
   </div>
