+++ spec/system/devise/sessions_spec.rb
@@ -2,7 +2,6 @@ require 'rails_helper'
 
 RSpec.describe "Devise/Sessions", type: :system do
   let!(:user) { create(:user, password: password) }
-  let(:user_with_facebook) { create(:user_with_facebook) }
   let(:password) { "password" }
 
   it "make session" do
@@ -27,24 +26,6 @@ RSpec.describe "Devise/Sessions", type: :system do
     end
     expect(page).to have_selector '.alert-success'
     expect(page).to have_current_path user_path(user)
-  end
-
-  it "make session with facebook" do
-    facebook_mock(user_with_facebook)
-    # ここまでの経路は"make sessionで確認済み"
-    visit new_user_session_path
-    click_on 'faceboook_login_button'
-    expect(page).to have_selector '.alert-success'
-    expect(page).to have_current_path user_path(user_with_facebook)
-  end
-
-  it "cannot make session with facebook" do
-    facebook_invalid_mock(user_with_facebook)
-    # ここまでの経路は"make user"で確認済み"
-    visit new_user_registration_path
-    click_on 'faceboook_login_button'
-    expect(page).to have_selector '.alert-warning'
-    expect(page).to have_current_path new_user_registration_path
   end
 
   it "delete session" do
